+++ backend/src/main/java/com/back/back9/global/jpa/entity/BaseEntity.java
@@ -1,6 +1,9 @@
 package com.back.back9.global.jpa.entity;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -12,11 +15,12 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
+@Getter
 public abstract class BaseEntity {
-
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private Long id;
+    @Setter(AccessLevel.PROTECTED)
+    Long id;
 
     @CreatedDate
     @Column(name = "created_at", updatable = false)
@@ -24,23 +28,7 @@ public abstract class BaseEntity {
 
     @LastModifiedDate
     @Column(name = "modified_at")
-    private LocalDateTime modifiedAt;
-
-    public Long getId() {
-        return id;
-    }
-
-    protected void setId(Long id) {
-        this.id = id;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public LocalDateTime getModifiedAt() {
-        return modifiedAt;
-    }
+    LocalDateTime modifiedAt;
 
     @Override
     public boolean equals(Object o) {
@@ -54,4 +42,56 @@ public abstract class BaseEntity {
     public int hashCode() {
         return Objects.hash(id);
     }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public LocalDateTime getModifiedAt() {
+        return modifiedAt;
+    }
 }
+
+
+//package com.back.back9.global.jpa.entity
+//
+//import jakarta.persistence.*
+//import org.springframework.data.annotation.CreatedDate
+//import org.springframework.data.annotation.LastModifiedDate
+//import org.springframework.data.jpa.domain.support.AuditingEntityListener
+//import java.time.LocalDateTime
+//
+//@MappedSuperclass
+//@EntityListeners(AuditingEntityListener::class)
+//abstract class BaseEntity {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    var id: Long? = null
+//        protected set
+//
+//    @CreatedDate
+//    @Column(name = "created_at", updatable = false)
+//    var createdAt: LocalDateTime? = null
+//        protected set
+//
+//
+//    @LastModifiedDate
+//    @Column(name = "modified_at")
+//    var modifiedAt: LocalDateTime? = null
+//        protected set
+//
+//    override fun equals(other: Any?): Boolean {
+//        if (this === other) return true
+//        if (other !is BaseEntity) return false
+//        return id != null && id == other.id
+//    }
+//
+//    override fun hashCode(): Int {
+//        return id?.hashCode() ?: 0
+//    }
+//}
