+++ backend/src/main/java/com/back/back9/global/security/SecurityUser.java
@@ -1,65 +1,65 @@
-//package com.back.back9.global.security;
-//
-//import com.back.back9.domain.user.entity.User;
-//import lombok.Getter;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.authority.SimpleGrantedAuthority;
-//import org.springframework.security.core.userdetails.UserDetails;
-//import org.springframework.security.oauth2.core.user.OAuth2User;
-//
-//import java.util.Collection;
-//import java.util.List;
-//import java.util.Map;
-//
-//@Getter
-//public class SecurityUser implements UserDetails, OAuth2User {
-//
-//    private final User user;
-//
-//    public SecurityUser(User user) {
-//        this.user = user;
-//    }
-//
-//    @Override
-//    public Collection<? extends GrantedAuthority> getAuthorities() {
-//        return List.of(new SimpleGrantedAuthority("ROLE_" + user.getRole().name()));
-//    }
-//
-//    @Override
-//    public String getPassword() {
-//        return user.getPassword();
-//    }
-//
-//    @Override
-//    public String getUsername() {
-//        return user.getUserLoginId();
-//    }
-//
-//    @Override
-//    public boolean isAccountNonExpired() { return true; }
-//
-//    @Override
-//    public boolean isAccountNonLocked() { return true; }
-//
-//    @Override
-//    public boolean isCredentialsNonExpired() { return true; }
-//
-//    @Override
-//    public boolean isEnabled() { return true; }
-//
-//    // OAuth2User 구현
-//    @Override
-//    public Map<String, Object> getAttributes() {
-//        return Map.of(
-//                "id", user.getId(),
-//                "userLoginId", user.getUserLoginId(),
-//                "username", user.getUsername(),
-//                "role", user.getRole().name()
-//        );
-//    }
-//
-//    @Override
-//    public String getName() {
-//        return user.getUserLoginId();
-//    }
-//}
+package com.back.back9.global.security;
+
+import com.back.back9.domain.user.entity.User;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+
+@Getter
+public class SecurityUser implements UserDetails, OAuth2User {
+
+    private final User user;
+
+    public SecurityUser(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority("ROLE_" + user.getRole().name()));
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getUserLoginId();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() { return true; }
+
+    @Override
+    public boolean isAccountNonLocked() { return true; }
+
+    @Override
+    public boolean isCredentialsNonExpired() { return true; }
+
+    @Override
+    public boolean isEnabled() { return true; }
+
+    // OAuth2User 구현
+    @Override
+    public Map<String, Object> getAttributes() {
+        return Map.of(
+                "id", user.getId(),
+                "userLoginId", user.getUserLoginId(),
+                "username", user.getUsername(),
+                "role", user.getRole().name()
+        );
+    }
+
+    @Override
+    public String getName() {
+        return user.getUserLoginId();
+    }
+}
