--- backend/src/main/kotlin/com/back/back9/global/jpa/entity/BaseEntity.kt
@@ -1,38 +1,0 @@
-package com.back.back9.global.jpa.entity
-
-import jakarta.persistence.*
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-abstract class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-        protected set
-
-    @CreatedDate
-    @Column(name = "created_at", updatable = false)
-    var createdAt: LocalDateTime? = null
-        protected set
-
-
-    @LastModifiedDate
-    @Column(name = "modified_at")
-    var modifiedAt: LocalDateTime? = null
-        protected set
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other !is BaseEntity) return false
-        return id != null && id == other.id
-    }
-
-    override fun hashCode(): Int {
-        return id?.hashCode() ?: 0
-    }
-}
