--- backend/src/main/kotlin/com/back/back9/global/webMvc/WebMvcConfig.java
@@ -1,19 +1,0 @@
-package com.back.back9.global.webMvc;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry
-                .addMapping("/**")  // 모든 경로에 대해 CORS 허용
-                .allowedOrigins("https://cdpn.io", "http://localhost:8888","https://peuronteuendeu.onrender.com","http://localhost:3000")
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
-                .allowedHeaders("*")
-                .allowCredentials(true)
-                .maxAge(3600);  // preflight 캐시 시간 설정
-    }
-}
